fix: hoist the subpath of sparse checkout
so the subpath is the root directory.

This ensures the subpath has its dependencies installed. Since cloned
repos are not currently crawled for package.json files, the package.json
has to exist in the root directory.

Note: This is not a long-term solution, since it breaks git pull and leaves the repo in an uncommitted state. Ideally, the repo would be in a hidden directory, and we would link to the subpath for the dest path.